crc_seed: u32

+++ src/crc.rs
@@ -4,7 +4,7 @@ use byteorder::{BigEndian, WriteBytesExt};
 use crc_table::get_crc_table;
 use wasm_bindgen::prelude::*;
 
-pub fn append_crc(data: &mut Vec<u8>, crc_seed: u8) -> () {
+pub fn append_crc(data: &mut Vec<u8>, crc_seed: u32) -> () {
     let crc = crc32(&data, (crc_seed >> 0) as usize);
     data.write_u16::<BigEndian>((crc & 0xffff) as u16).unwrap();
 }
